Split `document-mutator` into focused domain modules behind a barrel
